make a new topology library for topology operations shared between experiment and process
